adicionando e configurando models, seeders e factories

+++ database/migrations/2022_05_14_143127_create_addresses_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('addresses', function (Blueprint $table) {
             $table->id();
             $table->string('street');
-            $table->smallInteger('number');
+            $table->integer('number');
             $table->string('city');
             $table->string('district');
             $table->string('country');
